New ResourceBuilder Class implemented
